+++ apps/gotask_api/src/domain/interface/access/accessInterface.ts
@@ -1,0 +1,10 @@
+import { Document } from "mongoose";
+
+export interface IAccess extends Document {
+  id: string;
+  name: string;
+  application: {
+    access: string;
+    actions: string[];
+  }[];
+}
