+++ apps/gotask_api/src/domain/model/role.ts
@@ -1,12 +1,7 @@
-import { Document, Schema, model } from "mongoose";
+import { Schema, model } from "mongoose";
 import { v4 as uuidv4 } from "uuid";
+import { IRole } from "../interface/role/roleInterface";
 
-export interface IRole extends Document {
-  id: string;
-  name: string;
-  priority: number;
-  access: string[];
-}
 
 const RoleSchema = new Schema<IRole>(
   {
